refactor: Extract date formatting from cache.ts
This is in preparation for a fixup-at-startup step which will need to
log the files that were touched, and to that end requires date
formatting.
